api/backend/requirements.txt
flask==1.0.2
flask-sqlalchemy==2.3.1
pytest-flask==0.14.0
freezegun==0.1.11
delorean==1.0.0
faker==1.0.7
pyjwt==1.7.1
cryptography==2.6.1
parse==1.12.0
psycopg2==2.9.1
Werkzeug==0.16.1
bcrypt==3.2.0
boto3==1.16.53
tldextract==3.1.0
flask_restx==0.4.0